# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "O'graph",
        "description": "O'graph API" 
    },
    "host": "localhost:3000",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/usersByRole/{role}": {
            "parameters": [
                {
                    "name": "role",
                    "in": "path",
                    "required": true,
                    "description": "role of the user's",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users with given role",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/userById/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "id of the user's",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users with given id",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/usersByProjectId/{projectId}": {
            "parameters": [
                {
                    "name": "projectId",
                    "in": "path",
                    "required": true,
                    "description": "id of the project",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users with given project id",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/createUser": {
            "parameters": [
                {
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "user info"
                }
            ],
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create new User",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/updateUserPassword": {
            "parameters": [
                {
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "new user password"
                }
            ],
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user password",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/updateUser": {
            "parameters": [
                {
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "new user info"
                }
            ],
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user info",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}